Fix: totalPrice 업데이트 디버깅
