+++ apps/client/src/app/cart/page.tsx
@@ -2,8 +2,7 @@ import React from "react"
 import type { CartItemType } from "@/types/cart/cartTypes"
 import { getCartData } from "@/action/cart/cartAction"
 import { calculateTotalPrice } from "@/action/cart/cartDataAction"
-import CartItemContainer from "@/components/cart/organism/CartItemContainer"
-import CartCheckout from "@/components/cart/organism/CartCheckout"
+import CartTemplate from "@/components/cart/template/CartTemplate"
 
 async function cart() {
 	const items: CartItemType[] = await getCartData()
@@ -12,20 +11,11 @@ async function cart() {
 
 	return (
 		<main className="min-h-screen bg-black pt-10">
-			<div className="container mx-auto max-w-7xl">
-				<h1 className="mb-8 text-2xl font-bold text-white">My Cart</h1>
-				<div className="grid grid-cols-1 gap-8 lg:grid-cols-4">
-					<div className="lg:col-span-2">
-						<CartItemContainer
-							initialItems={items}
-							initialSelectedItems={selectedItem}
-						/>
-					</div>
-					<div className="h-auto">
-						<CartCheckout totalPrice={selectedItemsTotalPrice} />
-					</div>
-				</div>
-			</div>
+			<CartTemplate
+				initialItems={items}
+				initialSelectedItems={selectedItem}
+				initialTotalPrice={selectedItemsTotalPrice}
+			/>
 		</main>
 	)
 }
